+++ src/custom_flags.rs
@@ -7,6 +7,7 @@ use std::{
 
 use crate::{build_manager::BuildManager, per_test_config::TestConfig, Config, Errored};
 
+#[cfg(feature = "rustc")]
 pub mod run;
 pub mod rustfix;