+++ tests/integrations/basic/tests/ui_tests.rs
@@ -1,9 +1,8 @@
-use ui_test::*;
+use ui_test::{dependencies::DependencyBuilder, spanned::Spanned, *};
 
 fn main() -> ui_test::color_eyre::Result<()> {
     let path = "../../../target";
     let mut config = Config {
-        dependencies_crate_manifest_path: Some("Cargo.toml".into()),
         output_conflict_handling: if std::env::var_os("BLESS").is_some() {
             OutputConflictHandling::Bless
         } else {
@@ -16,6 +15,10 @@ fn main() -> ui_test::color_eyre::Result<()> {
     config.stdout_filter("in ([0-9]m )?[0-9\\.]+s", "");
     config.stderr_filter(r"[^ ]*/\.?cargo/registry/.*/", "$$CARGO_REGISTRY");
     config.path_stderr_filter(&std::path::Path::new(path), "$DIR");
+    config.comment_defaults.base().custom.insert(
+        "dependencies",
+        Spanned::dummy(vec![Box::new(DependencyBuilder::default())]),
+    );
 
     if let Ok(target) = std::env::var("UITEST_TEST_TARGET") {
         config.target = Some(target);